+++ nh_activity/__init__.py
@@ -1,4 +1,4 @@
 # Part of NHClinical. See LICENSE file for full copyright and licensing details
 # -*- coding: utf-8 -*-
-from . import activity
+from .models import *
 from .tests import test_model
