+++ nh_activity/models/__init__.py
@@ -1,0 +1,2 @@
+from . import activity
+from . import activity_data
