+++ nh_activity/tests/nh_activity/__init__.py
@@ -1,4 +1,5 @@
-from . import test_assign
+from .unit import *
+from .integration import *
 # from . import test_cancel
 # from . import test_complete
 # from . import test_create
