+++ nh_activity/tests/nh_activity/integration/__init__.py
@@ -1,0 +1,3 @@
+from . import test_activity_data_assign
+from . import test_activity_data_cancel
+from . import test_activity_data_complete
