+++ nh_activity/tests/nh_activity_data/__init__.py
@@ -1,1 +1,2 @@
-from . import test_activity_data_assign
+from data.nhclinical.nh_activity.tests.nh_activity.integration import \
+    test_activity_data_assign
